动态寻路

+++ GameServerLib/GameObjects/AttackableUnits/AI/ObjAIBase.cs
@@ -88,7 +88,7 @@ namespace LeagueSandbox.GameServer.GameObjects.AttackableUnits.AI
         public ICharScript CharScript { get; private set; }
         public bool IsBot { get; set; }
         public IAIScript AIScript { get; protected set; }
-        public ObjAIBase(Game game, string model, string name = "", int collisionRadius = 0,
+        public ObjAIBase(Game game, string model, string name = "", float collisionRadius = 0,
             Vector2 position = new Vector2(), int visionRadius = 0, int skinId = 0, uint netId = 0, TeamId team = TeamId.TEAM_NEUTRAL, Stats stats = null, string aiScript = "") :
             base(game, model, collisionRadius, position, visionRadius, netId, team, stats)
         {
